// client/stocklinx app/src/redux/fieldSet/constant.ts
export enum fieldSetConst {
  FETCH_FIELDSETS_REQUEST = "FETCH_FIELDSETS_REQUEST",
  FETCH_FIELDSETS_SUCCESS = "FETCH_FIELDSETS_SUCCESS",
  FETCH_FIELDSETS_FAILURE = "FETCH_FIELDSETS_FAILURE",
  FETCH_FIELDSET_REQUEST = "FETCH_FIELDSET_REQUEST",
  FETCH_FIELDSET_SUCCESS = "FETCH_FIELDSET_SUCCESS",
  FETCH_FIELDSET_FAILURE = "FETCH_FIELDSET_FAILURE",
  FETCH_FIELDSET_COUNTS_REQUEST = "FETCH_FIELDSET_COUNTS_REQUEST",
  FETCH_FIELDSET_COUNTS_SUCCESS = "FETCH_FIELDSET_COUNTS_SUCCESS",
  FETCH_FIELDSET_COUNTS_FAILURE = "FETCH_FIELDSET_COUNTS_FAILURE",
  CREATE_FIELDSET_REQUEST = "CREATE_FIELDSET_REQUEST",
  CREATE_FIELDSET_SUCCESS = "CREATE_FIELDSET_SUCCESS",
  CREATE_FIELDSET_FAILURE = "CREATE_FIELDSET_FAILURE",
  CREATE_RANGE_FIELDSET_REQUEST = "CREATE_RANGE_FIELDSET_REQUEST",
  CREATE_RANGE_FIELDSET_SUCCESS = "CREATE_RANGE_FIELDSET_SUCCESS",
  CREATE_RANGE_FIELDSET_FAILURE = "CREATE_RANGE_FIELDSET_FAILURE",
  UPDATE_FIELDSET_REQUEST = "UPDATE_FIELDSET_REQUEST",
  UPDATE_FIELDSET_SUCCESS = "UPDATE_FIELDSET_SUCCESS",
  UPDATE_FIELDSET_FAILURE = "UPDATE_FIELDSET_FAILURE",
  REMOVE_FIELDSET_REQUEST = "REMOVE_FIELDSET_REQUEST",
  REMOVE_FIELDSET_SUCCESS = "REMOVE_FIELDSET_SUCCESS",
  REMOVE_FIELDSET_FAILURE = "REMOVE_FIELDSET_FAILURE",
  REMOVE_RANGE_FIELDSET_REQUEST = "REMOVE_RANGE_FIELDSET_REQUEST",
  REMOVE_RANGE_FIELDSET_SUCCESS = "REMOVE_RANGE_FIELDSET_SUCCESS",
  REMOVE_RANGE_FIELDSET_FAILURE = "REMOVE_RANGE_FIELDSET_FAILURE",
  SET_FIELDSET = "SET_FIELDSET",
  CLEAR_FIELDSET = "CLEAR_FIELDSET",
  SET_FIELDSETS = "SET_FIELDSETS",
  CLEAR_FIELDSETS = "CLEAR_FIELDSETS",
}